Refine record sport routing and validation
